add concurrency function for expire, exists, ttl, pttl
